--- a/cpp_test_suite/environment/pre_test.sh.cmake
+++ b/cpp_test_suite/environment/pre_test.sh.cmake
@@ -9,11 +9,14 @@ echo "Run conf_devtest"
 echo "Start DevTest"
 echo "DevTest libraries:"
 echo "`ldd @PROJECT_BINARY_DIR@/cpp_test_ds/DevTest`"
-#TODO Starter?
 @PROJECT_BINARY_DIR@/cpp_test_ds/DevTest @INST_NAME@ -v5 1>@PROJECT_BINARY_DIR@/cpp_test_ds/DevTest_@INST_NAME@.out 2>&1 &
 echo $! > @PROJECT_BINARY_DIR@/cpp_test_ds/DevTest_@INST_NAME@.pid
 
-sleep 3
+if hash tango_admin 2>/dev/null; then
+    tango_admin --ping-device dserver/devtest/@INST_NAME@ 3
+else
+    sleep 3
+fi
 
 echo "Start FwdTest"
 echo "FwdTest libraries:"
@@ -21,4 +24,8 @@ echo "`ldd @PROJECT_BINARY_DIR@/cpp_test_ds/fwd_ds/FwdTest`"
 @PROJECT_BINARY_DIR@/cpp_test_ds/fwd_ds/FwdTest @INST_NAME@ -v5 1>@PROJECT_BINARY_DIR@/cpp_test_ds/fwd_ds/FwdTest_@INST_NAME@.out 2>&1 &
 echo $! > @PROJECT_BINARY_DIR@/cpp_test_ds/fwd_ds/FwdTest_@INST_NAME@.pid
 
-sleep 3
+if hash tango_admin 2>/dev/null; then
+    tango_admin --ping-device dserver/fwdtest/@INST_NAME@ 3
+else
+    sleep 3
+fi