… cpp_test_suite/environment/shutdown.sh