I propose to remove `-p` option when starting `tango_cs` container for tests.

It is not necessary to publish the port on host. We source generated `cpp_test_suite/tango_host` file, which specifies `TANGO_HOST` and uses container IP address directly.

With `-p` option, if you have tango already running on your system on port 10000, docker will refuse to start:
```
docker: Error response from daemon: driver failed programming external connectivity on endpoint pedantic_morse (b9bb28fc27ff178de0472035e0840bfda6c1291577e360013f4730dd8c3a9f5e): Error starting userland proxy: listen tcp 0.0.0.0:10000: bind: address already in use.
```