@Ingvord 2019-06-01
I use this script not only for tests but to run Tango for instance to test REST. It is more convenient than just raw Tango docker stack as this one has more devices (DevTest + FwdTest etc). So I would leave `-p` as it is. Is it that critical to run these scripts on machines with Tango installed, I mean it is intended to use it on a raw machine, like Travis. For developer it is also better to start fresh Tango to reduce effect of the persistent properties etc which may affect test results i.e. less environment depended the whole infrastructure - it is easier to reproduce and transfer it. Just my IMHO