@mliszcz 2019-06-01
Thanks! I think we can support REST testing usecase with something like:

```bash
PUBLISH_PORT="-p 10000:10000"
if lsof -tnPi TCP:10000 >/dev/null; then
    echo "Some process is already listening on port 10000, it will not be published"
    PUBLISH_PORT=""
fi
docker run $PUBLISH_PORT ... tango_cs
```

There is no technical reason to keep `-p` as tests are working without it. Let's give users freedom to configure (mess up) their system as they see fit.

I can put a note in `INSTALL.md` in test section that "it is recomended to shut down existing Tango instances before running the tests". (recommended != mandatory).