@Ingvord 2019-06-01
@mliszcz , you are right in the first place. After carefully rethinking I would say it is better to remove this port, as this will also allow to run multiple Tango docker stacks (if we name them properly) in the future. This will allow us to test multiple Tango hosts devices communication etc.

I have also checked if REST can access Tango via IP and this works fine (I was not sure about dots).

**BTW** do you know if we can set Starter name properly e.g. to IP or define a host name for docker? Now it is some rubbish, like `1aef7207caf2` in the picture belo:.

![rest](https://user-images.githubusercontent.com/11678364/58745720-2e974000-845d-11e9-9ce1-a4fcf29ad2b9.png)