@mliszcz 2019-06-01
`1aef7207caf2` is not rubbish but random (?) container hostname. You can provide your own hostname with `-h` (or `--hostname`). I can add it to this PR if you want, just tell me what value shall we use there.
https://docs.docker.com/config/containers/container-networking/#dns-services

If we are to touch that file, there is one more thing we can improve:
https://github.com/tango-controls/cppTango/blob/ab05abc289d187130614966e4c79e5390ca41b16/cpp_test_suite/environment/setup.sh#L33-L35
We could probably use:
```
tango_admin --ping-device sys/database/2 2
```
or
```
tango_admin --ping-database 2
```

The timeout (2 seconds in above examples) works in some unexpected way if the system is down:
```
$ time tango_admin --ping-device sys/database/2 1

real    0m12,642s
user    0m0,009s
sys     0m0,009s
$ time tango_admin --ping-device sys/database/2 2

real    0m25,807s
user    0m0,000s
sys     0m0,019s
$ time tango_admin --ping-device sys/database/2 3

real    0m38,948s
user    0m0,000s
sys     0m0,024s
$ time tango_admin --ping-device sys/database/2

real    0m6,101s
user    0m0,008s
sys     0m0,008s
```