@Ingvord 2019-06-01
> 1aef7207caf2 is not rubbish but random (?) container hostname. You can provide your own hostname with -h (or --hostname). I can add it to this PR if you want, just tell me what value shall we use there.
https://docs.docker.com/config/containers/container-networking/#dns-services

I have played a little bit with `-h|--hostname` and it does not work for me :( I have tried with `--hostname tango` and REST does not see this host:

![rest-resolve-tango-host-failure](https://user-images.githubusercontent.com/11678364/58747190-6360c280-8470-11e9-8226-3ffbf2fc082f.png)
 
Starter is named correctly though if Tango host is accessed via IP:

![rest-resolve-tango-host-failure1](https://user-images.githubusercontent.com/11678364/58747284-bd15bc80-8471-11e9-92be-a6fd4711c5f6.png)

And another problem is that devices are no longer accessible, as they point to `localhost:10000`

![rest-resolve-tango-host-failure2](https://user-images.githubusercontent.com/11678364/58747305-eafb0100-8471-11e9-8dd1-a1d6b6266496.png)

So my conclusion is that we need more careful setup for Tango docker stack, to address all the issues above... 

Let's fallback for `if-fi` solution for the moment.

`tango-admin ping` seems like a good improvement to me!