@mliszcz 2019-06-01
> I have played a little bit with -h|--hostname and it does not work for me :( I have tried with --hostname tango and REST does not see this host:

Because REST server is running on localhost (outside docker) and hostname resolution will not work in this case. `-h` will only set hostname inside docker container (and will add entry to docker's internal DNS so that different containers could talk to each other by hostname). To access container by its hostname from your machine you need to modify your system configuration with either:
* additional DNS proxy server shared by host and containers or
* hardcode container's IP address in `/etc/hosts` file

> Starter is named correctly though if Tango host is accessed via IP:

This is expected as hostname is changed in starter's container, and starter server is started with:
```
/usr/bin/Starter `hostname --short`
```

> And another problem is that devices are no longer accessible, as they point to localhost:10000 (Why?!)

This is probably configuration issue. Is correct `TANGO_HOST` set for the devices? Maybe they are reading `/etc/tangorc` or are using some default value.

> So my conclusion is that we need more careful setup for Tango docker stack, to address all the issues above...

There were some efforts in latest TangoBox release to dockerize as much as possible. If there is a demand (and proper priotity is assiged), we can implement DNS resolution and move DataBaseds to a container. I did some tests few weeks ago and it is possible to expose docker's internal DNS server to the host, but some effort is still required to make it usable.