@mliszcz 2019-06-05
I've did some more changes in latest commit:
* Remove -p, so that another instance can use port 10000
* Speedup DB setup by disabling InnoDB
* Use tango_admin to wait until the DB goes online

Results:

Before the PR - we sleep for 30s:
```
$ time make start-tango
...
real    0m32,212s
user    0m0,217s
sys     0m0,123s
```

With --ping-database instead of sleep - 15s to start the db:
```
real    0m16,085s
user    0m0,364s
sys     0m0,206s
```

With InnoDB and timezone detection disabled - just 5 seconds to start the db:
```
real    0m5,367s
user    0m0,216s
sys     0m0,155s
```

I think we need this kind of improvement anyway if we want to re-start the DB and servers before each test.