@mliszcz 2019-06-05
Also changed pre_test.sh to use tango_admin:
* Use tango_admin to wait until device servers are online