@mliszcz 2019-06-16
Obsoleted by #568