@mliszcz 2019-11-26
Closing in favor of #568